% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelFeatsFunctions.R
\name{labelSingleFeat}
\alias{labelSingleFeat}
\title{Label a single chromatographic feature}
\usage{
labelSingleFeat(
  feature_data_i,
  ms1_data,
  ppm_window_width = 10,
  rt_window_width = 2
)
}
\arguments{
\item{feature_data_i}{Information associated with a single chromatographic
feature, most importantly its central *m/z* value (mzmed) and retention time
in minutes (rtmed).}

\item{ms1_data}{A data.table containing mz, rt, and int columns containing
the MS1 data from which the feature's chromatogram should be extracted.}

\item{ppm_window_width}{The parts-per-million window of the extracted ion chromatogram}

\item{rt_window_width}{The width of the retention time window for the extracted
ion chromatogram, in minutes.}
}
\value{
A scalar character corresponding to either "Quit", "Backspace",
"Good", "Bad", or "Revisit", depending on what key is pressed while the plot
window is active.
}
\description{
This function is the core operation of the labelFeatsManual function.
The plot appears in a new window and is set up to detect key presses which
have been bound to specific classifications. Currently, the left arrow key
categorizes the feature as "Bad" while the right arrow key categorizes it
as "Good". Up and down arrows can be bound to additional classifications if
desired by editing the code, though a future update may provide more control.
}
\examples{
library(RaMS)
mzML_files <- system.file("extdata", package = "RaMS") |>
    list.files(full.names = TRUE, pattern = "[A-F].mzML")
msdata <- grabMSdata(mzML_files, verbosity = 0)

if (interactive()) {
    labelSingleFeat(data.frame(mzmed = 118.0865, rtmed = 7.8), ms1_data = msdata$MS1)
}

# Returns "Bad" if the extracted ion chromatogram has fewer than 5 scans
labelSingleFeat(data.frame(mzmed = 50.000, rtmed = 5),
    ms1_data = msdata$MS1,
    ppm_window_width = 1, rt_window_width = 10
)
}
