% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{makeXcmsObjFlat}
\alias{makeXcmsObjFlat}
\title{Make an XCMS object flat}
\usage{
makeXcmsObjFlat(xcms_obj, revert_rts = TRUE)
}
\arguments{
\item{xcms_obj}{A XCMSnExp object produced by the typical XCMS workflow which
should include retention time correction and peak correspondence and filling}

\item{revert_rts}{Scalar boolean controlling whether the adjusted retention
times found in the XCMS object are propagated or returned as-is}
}
\value{
A data.frame with columns for feature information (from 
featureDefinitions: feature, feat_mzmed, feat_rtmed, feat_npeaks, and
peakidx) and peak information (from chromPeaks: mz, mzmin, mzmax, rt, rtmin,
rtmax, into, intb, maxo, sn, sample) as well as the full path to the 
associated file and the file name alone (filepath and filename).
}
\description{
XCMSnExp objects are complicated S4 objects that make
it difficult to access information about the features and their associated
peaks. This function turns the output into a "flat" file format (a 
data.frame) so it can interact with tidyverse functions more easily.
}
\examples{
msnexp_filled <- readRDS(system.file("extdata", "intro_xcms_filled.rds", package="squallms"))
makeXcmsObjFlat(msnexp_filled)
}
