% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelFeatsFunctions.R
\name{labelFeatsManual}
\alias{labelFeatsManual}
\title{Label chromatographic features manually}
\usage{
labelFeatsManual(
  peak_data,
  ms1_data = NULL,
  existing_labels = NULL,
  selection = "Unlabeled",
  verbosity = 1
)
}
\arguments{
\item{peak_data}{Flat-form XC-MS data with columns for the bounding box of
a chromatographic peak (mzmin, mzmax, rtmin, rtmax) as grouped by a feature 
ID. Must be provided WITHOUT retention time correction for proper matching
to the values in the raw data.}

\item{ms1_data}{Optional data.table object produced by RaMS containing MS1
data with columns for filename, rt, mz, and int. If not provided, the files
are detected from the filepath column in peak_data.}

\item{existing_labels}{A character vector of equal length to the number of
features in peak_data named with feature IDs. Can be used to provide a
previously-existing partially-labeled dataset or double-check existing
classifications.}

\item{selection}{Either the string "Labeled" or "Unlabeled". If "Unlabeled",
the classifier will target entries in the dataset that have not yet received
a classification (i.e. those that have NA values in existing_labels). If
"Labeled", the classifier will target (and overwrite) existing labels. 
Otherwise, the classifier will randomly target any feature whether previously
classified or not.}
}
\value{
A character vector named with feature IDs containing the classifications
of each peak that was viewed during the interactive phase. NA values indicate
those features that were not classified.
}
\description{
This function provides an interactive interface to a labeling function that
plots one chromatographic feature at a time, showing the data from all the
files in ms1_data within 1 minute of retention time and 10 ppm m/z space.
The plot appears in a new window and is set up to detect key presses which
have been bound to specific classifications. Currently, the left arrow key
categorizes the feature as "Bad" while the right arrow key categorizes it
as "Good". Up and down arrows can be bound to additional classifications if
desired by editing the code, though a future update may provide more control.
Classifications are returned as a character vector named with the feature
IDs even if the full dataset is not classified. Exit the classifier with
the Escape key.
}
\examples{
msnexp_filled <- readRDS("vignettes/intro_xcms_filled.rds")
peak_data <- makeXcmsObjFlat(msnexp_filled)
manual_labels <- labelFeatsManual(peak_data)
}
