% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateXcmsObjFeatsFunctions.R
\name{logModelFeatProb}
\alias{logModelFeatProb}
\title{Title}
\usage{
logModelFeatProb(
  feature_metrics,
  feature_labels,
  log_formula = feat_class ~ med_cor + med_snr,
  verbosity = 2
)
}
\arguments{
\item{feature_metrics}{A data.frame with columns used to construct the 
feature quality model.}

\item{feature_labels}{A character vector named with feature IDs and entries
corresponding to the peak quality (either "Good", "Bad", or NA).}

\item{log_formula}{The formula to use when predicting feature quality from
the feature metrics. This formula is passed to `glm` as-is, so make sure that
the predictive features exist in the feature_metrics data.frame.}

\item{verbosity}{Scalar value between zero and two determining how much 
diagnostic information is produced. Zero should return nothing, one should
return text-based progress markers, and 2 will return diagnostic plots.}
}
\description{
Title
}
